backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
google-generativeai==0.3.2
gspread==5.12.0
google-auth==2.25.2
oauth2client==4.1.3
websockets==12.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
requests==2.31.0
aiofiles==23.2.1
motor==3.3.2